Add JMX and REST support for diagnostic events  See https://issues.apache.org/jira/browse/CASSANDRA-12944 – Diagnostic Events
ref: https://github.com/thelastpickle/cassandra-reaper/pull/441

+++ src/server/src/main/java/io/cassandrareaper/resources/ClusterResource.java
@@ -357,6 +357,11 @@ public final class ClusterResource {
           .entity("cluster \"" + clusterName + "\" cannot be deleted, as it has repair runs")
           .build();
     }
+    if (!context.storage.getEventSubscriptions(clusterName).isEmpty()) {
+      return Response.status(Response.Status.CONFLICT)
+              .entity("cluster \"" + clusterName + "\" cannot be deleted, as it has diagnostic events subscriptions")
+              .build();
+    }
     context.storage.deleteCluster(clusterName);
     return Response.accepted().build();
   }
